+++ QATLibraryTests/QATQuestionsDataSourceTests.m
@@ -39,7 +39,7 @@
     NSMutableArray* json_object = [NSMutableArray arrayWithCapacity:3];
     
     for (NSInteger index=0; index<numberOfObjects; index++) {
-        NSDictionary * dict = [NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"item%d",index],@"content", nil];
+        NSDictionary * dict = [NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"item%ld",(long)index],@"content", nil];
         [json_object addObject:dict];
     }
     
@@ -69,7 +69,7 @@
 {
     QATQuestionsDataSource *questions = [[QATQuestionsDataSource alloc] initWithConnection:self.doesNotMatter];
     
-    XCTAssertEqual(questions.length, 0,@"Before loading any data the length should be 0!");
+    XCTAssertEqual((int)questions.length, 0,@"Before loading any data the length should be 0!");
     
 }
 
@@ -93,7 +93,7 @@
     
     [questions downloadCompleted:[self createJSONDataFromJSONObject:[self generateTestJSONObjectWith:5]]];
     
-    XCTAssertEqual(5, questions.length,@"Incorrect number of objects returned.");
+    XCTAssertEqual(5, (int)questions.length,@"Incorrect number of objects returned.");
 }
 
 - (void)testReceivedQuestionsAreTheSameAsTheQuestionsSentFromTheServer
